infinitedsp_core::core::ola

Trait SpectralProcessor

pub trait SpectralProcessor {
    // Required method
    fn process_spectral(&mut self, bins: &mut [Complex32], sample_index: u64);

    // Provided method
    fn name(&self) -> &str { ... }
}

Trait for processors that operate on spectral data (FFT bins).

Required Methods

fn process_spectral(&mut self, bins: &mut [Complex32], sample_index: u64)

Process a block of complex spectral bins.

Arguments

bins - The spectral data.

sample_index - The sample index corresponding to the start of the analysis window.

Provided Methods

fn name(&self) -> &str

Returns the name of the spectral processor.