infinitedsp_core::effects::filter::vowel

Struct VowelFilter

pub struct VowelFilter {
    pub f1: StateVariableFilter,
    pub f2: StateVariableFilter,
    pub f3: StateVariableFilter,
    vowel_morph: AudioParam,
    q: AudioParam,
    sample_rate: f32,
    manual_formants: Option<(f32, f32, f32)>,
}

A Vowel (Formant) Filter.

Simulates the human vocal tract by using three parallel band-pass filters tuned to specific formant frequencies. It can morph between different vowels using the vowel_morph parameter (0.0 to 4.0).

Fields
f1: StateVariableFilter

Internal filter for F1.

f2: StateVariableFilter

Internal filter for F2.

f3: StateVariableFilter

Internal filter for F3.

vowel_morph: AudioParamq: AudioParamsample_rate: f32manual_formants: Option<(f32, f32, f32)>
Implementations

impl VowelFilter

pub const A: (f32, f32, f32)

Vowel ‘A’ formant frequencies.

pub const E: (f32, f32, f32)

Vowel ‘E’ formant frequencies.

pub const I: (f32, f32, f32)

Vowel ‘I’ formant frequencies.

pub const O: (f32, f32, f32)

Vowel ‘O’ formant frequencies.

pub const U: (f32, f32, f32)

Vowel ‘U’ formant frequencies.

pub const S: (f32, f32, f32)

Consonant ‘S’ formant frequencies.

pub const Z: (f32, f32, f32)

Consonant ‘Z’ formant frequencies.

pub const F: (f32, f32, f32)

Consonant ‘F’ formant frequencies.

pub const V: (f32, f32, f32)

Consonant ‘V’ formant frequencies.

pub const H: (f32, f32, f32)

Consonant ‘H’ formant frequencies.

pub const TH: (f32, f32, f32)

Consonant ‘TH’ formant frequencies.

pub const SH: (f32, f32, f32)

Consonant ‘SH’ formant frequencies.

pub const R: (f32, f32, f32)

Consonant ‘R’ formant frequencies.

pub const L: (f32, f32, f32)

Consonant ‘L’ formant frequencies.

pub const N: (f32, f32, f32)

Consonant ‘N’ formant frequencies.

pub const M: (f32, f32, f32)

Consonant ‘M’ formant frequencies.

pub const NG: (f32, f32, f32)

Consonant ‘NG’ formant frequencies.

pub const W: (f32, f32, f32)

Consonant ‘W’ formant frequencies.

pub const Y: (f32, f32, f32)

Consonant ‘Y’ formant frequencies.

pub fn new(vowel_morph: AudioParam, q: AudioParam) -> Self

Creates a new VowelFilter.

Arguments

vowel_morph - A parameter (0.0 to 4.0) to morph between A, E, I, O, U.

q - The resonance (Q) of the formant filters.

fn get_formants_at(&self, morph: f32) -> (f32, f32, f32)

pub fn set_formants(&mut self, f1: f32, f2: f32, f3: f32)

Manually sets the formant frequencies. This overrides the vowel_morph parameter.

pub fn set_q(&mut self, q: AudioParam)

Sets the resonance (Q) for all formant filters.

pub fn tick_manual( &mut self, input: f32, f1: f32, f2: f32, f3: f32, q: f32, ) -> f32

Efficiently processes a single sample with manual formant control.