infinitedsp_core::synthesis::envelope

Struct Adsr

pub struct Adsr {

    gate: AudioParam,
    attack_time: AudioParam,
    decay_time: AudioParam,
    sustain_level: AudioParam,
    release_time: AudioParam,
    sample_rate: f32,
    state: AdsrState,
    current_level: f32,
    last_gate: f32,
    attack_step: f32,
    decay_coeff: f32,
    release_coeff: f32,
    last_attack: f32,
    last_decay: f32,
    last_release: f32,
    gate_buffer: Vec<f32>,
    attack_buffer: Vec<f32>,
    decay_buffer: Vec<f32>,
    sustain_buffer: Vec<f32>,
    release_buffer: Vec<f32>,
    retrigger: Arc<AtomicBool>,

}

An ADSR (Attack, Decay, Sustain, Release) envelope generator.

Generates a control signal based on a gate input. Time parameters are in seconds.

Fields
gate: AudioParamattack_time: AudioParamdecay_time: AudioParamsustain_level: AudioParamrelease_time: AudioParamsample_rate: f32state: AdsrStatecurrent_level: f32last_gate: f32attack_step: f32decay_coeff: f32release_coeff: f32last_attack: f32last_decay: f32last_release: f32gate_buffer: Vec<f32>attack_buffer: Vec<f32>decay_buffer: Vec<f32>sustain_buffer: Vec<f32>release_buffer: Vec<f32>retrigger: Arc<AtomicBool>
Implementations

impl Adsr

pub fn new( gate: AudioParam, attack_time: AudioParam, decay_time: AudioParam, sustain_level: AudioParam, release_time: AudioParam, ) -> Self

Creates a new ADSR envelope.

Arguments

gate - Gate signal (0.0 = off, 1.0 = on).

attack_time - Attack time in seconds.

decay_time - Decay time in seconds.

sustain_level - Sustain level (0.0 - 1.0).

release_time - Release time in seconds.

pub fn create_trigger(&self) -> Trigger

Creates a trigger handle for this envelope. Use this to manually retrigger the envelope from any thread.

fn recalc(&mut self, attack: f32, decay: f32, release: f32)

pub fn set_attack(&mut self, time: AudioParam)

Sets the attack time parameter (seconds).

pub fn set_decay(&mut self, time: AudioParam)

Sets the decay time parameter (seconds).

pub fn set_sustain(&mut self, level: AudioParam)

Sets the sustain level parameter.

pub fn set_release(&mut self, time: AudioParam)

Sets the release time parameter (seconds).