infinitedsp_core::synthesis::speech

Struct SpeechSynth

pub struct SpeechSynth<'a> {

    sample_rate: f32,
    time_in_phoneme: f32,
    phonemes: &'a [Phoneme],
    current_idx: usize,
    cur_f1: f32,
    cur_f2: f32,
    cur_f3: f32,
    cur_voice: f32,
    cur_noise: f32,
    cur_pitch: f32,
    cur_amp: f32,
    vowel_filter: VowelFilter,
    stutter: Stutter,
    voice_stack: Stack,
    sub_osc: Oscillator,
    noise_gen: Oscillator,
    just_entered_phoneme: bool,
    is_finished: bool,

}

A complete Speech Synthesizer component.

Implements a formant-based vocal synthesizer that can play back sequences of phonemes.

Fields
sample_rate: f32time_in_phoneme: f32phonemes: &'a [Phoneme]current_idx: usizecur_f1: f32cur_f2: f32cur_f3: f32cur_voice: f32cur_noise: f32cur_pitch: f32cur_amp: f32vowel_filter: VowelFilterstutter: Stuttervoice_stack: Stacksub_osc: Oscillatornoise_gen: Oscillatorjust_entered_phoneme: boolis_finished: bool
Implementations

impl<'a> SpeechSynth<'a>

pub fn new(sample_rate: f32) -> Self

Creates a new SpeechSynth.

pub fn set_phonemes(&mut self, phonemes: &'a [Phoneme])

Sets the phoneme sequence to be played.

pub fn is_finished(&self) -> bool

Returns true if the phoneme sequence has finished playing.